cli: move the zip tests to their own file
Release note: None
